Move fields to own folder
