+++ src/ig-template/features/Feature.ts
@@ -2,7 +2,7 @@ import {Saveable} from "@/ig-template/tools/saving/Saveable";
 import {SaveData} from "@/ig-template/tools/saving/SaveData";
 import {CurrencyType} from "@/ig-template/features/wallet/CurrencyType";
 import {Features} from "@/ig-template/Features";
-import {AbstractField} from "@/ig-template/developer-panel/AbstractField";
+import {AbstractField} from "@/ig-template/developer-panel/fields/AbstractField";
 
 /**
  * An abstract class that all features should extend for.
