+++ src/ig-template/features/settings/Settings.ts
@@ -5,8 +5,8 @@ import {SettingId} from "@/ig-template/features/settings/SettingId";
 import {MultipleChoiceSetting} from "@/ig-template/features/settings/MultipleChoiceSetting";
 import {SettingOption} from "@/ig-template/features/settings/SettingOption";
 import {Features} from "@/ig-template/Features";
-import {AbstractField} from "@/ig-template/developer-panel/AbstractField";
-import {FunctionField} from "@/ig-template/developer-panel/FunctionField";
+import {AbstractField} from "@/ig-template/developer-panel/fields/AbstractField";
+import {FunctionField} from "@/ig-template/developer-panel/fields/FunctionField";
 
 export class Settings extends Feature {
     list: Setting<any>[];
